pyproject.toml
[tool.poetry]
name = "sbom-to-snyk-depgraph"
version = "0.1.0"
description = "convert sbom to snyk depgraph and test or monitor"
authors = ["Scott Esbrandt <scott.esbrandt@snyk.io>"]
license = "Apache-2.0"
readme = "README.md"
packages = [{include = "sbom_to_snyk_depgraph"}]

[tool.poetry.dependencies]
python = "^3.9"
pysnyk = "^0.9.18"
typer = "^0.9.0"
cyclonedx-python-lib = "^6.0.0"

[tool.poetry.group.dev.dependencies]
black = "^23.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
